src/requirements.txt
Flask==2.1.3
flask-blueprint==1.3.0
ipdb==0.13.9
requests==2.28.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.4
flower==1.1.0
Flask-CeleryExt==0.4.3